backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-genai==1.14.0
pydantic==2.10.0
httpx==0.28.0
python-dotenv==1.0.1
aiosqlite==0.21.0
databases==0.9.0
jinja2==3.1.4
yt-dlp==2025.3.31
beautifulsoup4==4.12.3
